Feat/popup (#68)
* fix: async function error

* feat: home page popups done reviewcards take you to users page
